+++ effectful-opaleye/CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Haskell Package Versioning Policy](https://pvp.hask
 
 ## [Unreleased]
 
+### Added
+
+- Ability to keep a running tally of the SQL operations that are performed by
+  the `Opaleye` effect in #4.
+
 ## [0.1.0.1] - 04.08.2025
 
 ### Changed
@@ -23,5 +28,5 @@ and this project adheres to [Haskell Package Versioning Policy](https://pvp.hask
 - CI that builds and tests the packages for each version of GHC in the `tested-with` field.
 
 [unreleased]: https://github.com/fpringle/effectful-postgresql/compare/v0.1.0.1...HEAD
-[0.1.0.1]: https://github.com/fpringle/effectful-postgresql/releases/tag/v0.1.0.1
+[0.1.0.1]: https://github.com/fpringle/effectful-postgresql/compare/v0.1.0.0...v0.1.0.1
 [0.1.0.0]: https://github.com/fpringle/effectful-postgresql/releases/tag/v0.1.0.0